+++ meta/types_gen.go
@@ -467,6 +467,227 @@ func (z *DDLJob) Msgsize() (s int) {
 }
 
 // DecodeMsg implements msgp.Decodable
+func (z *DropJob) DecodeMsg(dc *msgp.Reader) (err error) {
+	var field []byte
+	_ = field
+	var zb0001 uint32
+	zb0001, err = dc.ReadMapHeader()
+	if err != nil {
+		err = msgp.WrapError(err)
+		return
+	}
+	for zb0001 > 0 {
+		zb0001--
+		field, err = dc.ReadMapKeyPtr()
+		if err != nil {
+			err = msgp.WrapError(err)
+			return
+		}
+		switch msgp.UnsafeString(field) {
+		case "Table":
+			z.Table, err = dc.ReadString()
+			if err != nil {
+				err = msgp.WrapError(err, "Table")
+				return
+			}
+		case "Cursor":
+			z.Cursor, err = dc.ReadInt()
+			if err != nil {
+				err = msgp.WrapError(err, "Cursor")
+				return
+			}
+		case "Started":
+			z.Started, err = dc.ReadInt64()
+			if err != nil {
+				err = msgp.WrapError(err, "Started")
+				return
+			}
+		case "Def":
+			if dc.IsNil() {
+				err = dc.ReadNil()
+				if err != nil {
+					err = msgp.WrapError(err, "Def")
+					return
+				}
+				z.Def = nil
+			} else {
+				if z.Def == nil {
+					z.Def = new(TableDef)
+				}
+				err = z.Def.DecodeMsg(dc)
+				if err != nil {
+					err = msgp.WrapError(err, "Def")
+					return
+				}
+			}
+		default:
+			err = dc.Skip()
+			if err != nil {
+				err = msgp.WrapError(err)
+				return
+			}
+		}
+	}
+	return
+}
+
+// EncodeMsg implements msgp.Encodable
+func (z *DropJob) EncodeMsg(en *msgp.Writer) (err error) {
+	// map header, size 4
+	// write "Table"
+	err = en.Append(0x84, 0xa5, 0x54, 0x61, 0x62, 0x6c, 0x65)
+	if err != nil {
+		return
+	}
+	err = en.WriteString(z.Table)
+	if err != nil {
+		err = msgp.WrapError(err, "Table")
+		return
+	}
+	// write "Cursor"
+	err = en.Append(0xa6, 0x43, 0x75, 0x72, 0x73, 0x6f, 0x72)
+	if err != nil {
+		return
+	}
+	err = en.WriteInt(z.Cursor)
+	if err != nil {
+		err = msgp.WrapError(err, "Cursor")
+		return
+	}
+	// write "Started"
+	err = en.Append(0xa7, 0x53, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64)
+	if err != nil {
+		return
+	}
+	err = en.WriteInt64(z.Started)
+	if err != nil {
+		err = msgp.WrapError(err, "Started")
+		return
+	}
+	// write "Def"
+	err = en.Append(0xa3, 0x44, 0x65, 0x66)
+	if err != nil {
+		return
+	}
+	if z.Def == nil {
+		err = en.WriteNil()
+		if err != nil {
+			return
+		}
+	} else {
+		err = z.Def.EncodeMsg(en)
+		if err != nil {
+			err = msgp.WrapError(err, "Def")
+			return
+		}
+	}
+	return
+}
+
+// MarshalMsg implements msgp.Marshaler
+func (z *DropJob) MarshalMsg(b []byte) (o []byte, err error) {
+	o = msgp.Require(b, z.Msgsize())
+	// map header, size 4
+	// string "Table"
+	o = append(o, 0x84, 0xa5, 0x54, 0x61, 0x62, 0x6c, 0x65)
+	o = msgp.AppendString(o, z.Table)
+	// string "Cursor"
+	o = append(o, 0xa6, 0x43, 0x75, 0x72, 0x73, 0x6f, 0x72)
+	o = msgp.AppendInt(o, z.Cursor)
+	// string "Started"
+	o = append(o, 0xa7, 0x53, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64)
+	o = msgp.AppendInt64(o, z.Started)
+	// string "Def"
+	o = append(o, 0xa3, 0x44, 0x65, 0x66)
+	if z.Def == nil {
+		o = msgp.AppendNil(o)
+	} else {
+		o, err = z.Def.MarshalMsg(o)
+		if err != nil {
+			err = msgp.WrapError(err, "Def")
+			return
+		}
+	}
+	return
+}
+
+// UnmarshalMsg implements msgp.Unmarshaler
+func (z *DropJob) UnmarshalMsg(bts []byte) (o []byte, err error) {
+	var field []byte
+	_ = field
+	var zb0001 uint32
+	zb0001, bts, err = msgp.ReadMapHeaderBytes(bts)
+	if err != nil {
+		err = msgp.WrapError(err)
+		return
+	}
+	for zb0001 > 0 {
+		zb0001--
+		field, bts, err = msgp.ReadMapKeyZC(bts)
+		if err != nil {
+			err = msgp.WrapError(err)
+			return
+		}
+		switch msgp.UnsafeString(field) {
+		case "Table":
+			z.Table, bts, err = msgp.ReadStringBytes(bts)
+			if err != nil {
+				err = msgp.WrapError(err, "Table")
+				return
+			}
+		case "Cursor":
+			z.Cursor, bts, err = msgp.ReadIntBytes(bts)
+			if err != nil {
+				err = msgp.WrapError(err, "Cursor")
+				return
+			}
+		case "Started":
+			z.Started, bts, err = msgp.ReadInt64Bytes(bts)
+			if err != nil {
+				err = msgp.WrapError(err, "Started")
+				return
+			}
+		case "Def":
+			if msgp.IsNil(bts) {
+				bts, err = msgp.ReadNilBytes(bts)
+				if err != nil {
+					return
+				}
+				z.Def = nil
+			} else {
+				if z.Def == nil {
+					z.Def = new(TableDef)
+				}
+				bts, err = z.Def.UnmarshalMsg(bts)
+				if err != nil {
+					err = msgp.WrapError(err, "Def")
+					return
+				}
+			}
+		default:
+			bts, err = msgp.Skip(bts)
+			if err != nil {
+				err = msgp.WrapError(err)
+				return
+			}
+		}
+	}
+	o = bts
+	return
+}
+
+// Msgsize returns an upper bound estimate of the number of bytes occupied by the serialized message
+func (z *DropJob) Msgsize() (s int) {
+	s = 1 + 6 + msgp.StringPrefixSize + len(z.Table) + 7 + msgp.IntSize + 8 + msgp.Int64Size + 4
+	if z.Def == nil {
+		s += msgp.NilSize
+	} else {
+		s += z.Def.Msgsize()
+	}
+	return
+}
+
+// DecodeMsg implements msgp.Decodable
 func (z *IndexDef) DecodeMsg(dc *msgp.Reader) (err error) {
 	var field []byte
 	_ = field